apps/backends/stocks-backend/core/requirements/base.txt
celery==5.4.0
dj-database-url==2.3.0
Django==4.2
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-debug-toolbar==4.3.0
django-errformer==0.1.1
django-jazzmin==3.0.1
django-redis
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf_yasg==1.21.8
environs==11.2.0
gunicorn==23.0.0
httpx==0.28.1
pandas==2.2.3
redis==5.2.0
requests==2.32.3
psycopg2-binary==2.9.10
uvicorn==0.32.0
whitenoise==6.8.2
